backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyjwt==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
httpx==0.24.1
pytest==7.4.3
pytest-asyncio==0.21.1
loguru==0.7.2
supabase==2.0.2
gotrue==1.3.0
storage3==0.6.1
python-dateutil==2.8.2
email-validator==2.1.0
